+++ src/main/java/vexatos/backpacks/backpack/BackpackComputer.java
@@ -3,12 +3,21 @@ package vexatos.backpacks.backpack;
 import cpw.mods.fml.common.registry.GameRegistry;
 import vexatos.backpacks.reference.Mods;
 
+import static vexatos.backpacks.reference.Mods.*;
+
 /**
  * @author Vexatos
  */
 public class BackpackComputer extends BackpackModBase {
 	public BackpackComputer() {
-		super("computer", 0x333333, 0xF0F000, Mods.OpenComputers, Mods.ComputerCraft, Mods.Computronics);
+		super("computer", 0x333333, 0xF0F000,
+			ComputerCraft, OpenComputers, Computronics, OpenPeripheralAddons,
+
+			ComputerCraftEdu, RandomPeripherals, OpenCCSensors, CCLights2, PeripheralsPlusPlus, MoarPeripherals,
+			CCTweaks, LyqydPeripherals, ChunkyPeripherals, ImmibisPeripherals, FirePeripherals, DixPeripherals, Quadracopters,
+
+			OCLights2, OpenSecurity, OpenPrinter, OpenLights, OpenGlasses, MassSound, OpenAutomation
+		);
 	}
 
 	@Override