+++ changelog.md
@@ -1,3 +1,8 @@
+==== 1.0.1 ====
+
+  * ADDED: Flamingo Backpack, due to the great demand
+  * ADDED: Most ComputerCraft and OpenComputers addons are supported by the Computer Engineer's Backpack now. If there is any addon I missed, tell me.
+
 ==== 1.0.0 ====
 
   * Initial Release